move all build/dev related files to build-tools
