+++ .eslintignore
@@ -1,4 +1,3 @@
-build/*.js
-bin/*.js
-config/*.js
+build/**/*.js
+build-tools/**/*.js
 src/vendor/**/*.js
