+++ .gitignore
@@ -1,6 +1,6 @@
 .DS_Store
 node_modules/
-dist/
+build/
 .idea
 stats.json
 */.idea
