+++ .senggenerator
@@ -1,4 +1,4 @@
 {
-  "templatePath": "./template",
+  "templatePath": "./build-tools/template",
   "destination": "./src/component"
 }
