+++ build-tools/config/webpack/webpack.prod.conf.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const config = require('../config');
+const config = require('../../config');
 const webpack = require('webpack');
 const merge = require('webpack-merge');
 const baseWebpackConfig = require('./webpack.base.conf');
@@ -114,7 +114,7 @@ const webpackConfig = merge(baseWebpackConfig, {
         return (
           module.resource &&
           /\.js$/.test(module.resource) &&
-          module.resource.indexOf(path.join(__dirname, '../node_modules')) === 0
+          module.resource.indexOf(path.join(__dirname, '../../node_modules')) === 0
         );
       },
     }),
