+++ build-tools/template/complex-store/.senggenerator
