+++ build-tools/template/complex-store/{name_cc}/actions.js
