+++ build-tools/template/complex-store/{name_cc}/getters.js
