+++ build-tools/template/complex-store/{name_cc}/index.js
