+++ build-tools/template/complex-store/{name_cc}/state.js
