+++ build-tools/template/complex-store/{name_cc}/{name_cc}.js
