+++ build-tools/template/component/.senggenerator
