+++ build-tools/template/component/{name_pc}/index.js
