+++ build-tools/template/component/{name_pc}/{name_pc}.js
