+++ build-tools/template/page/.senggenerator
