+++ build-tools/template/page/{name_pc}/index.js
