+++ build-tools/template/page/{name_pc}/{name_pc}.scss
