+++ build-tools/template/store/.senggenerator
