+++ build-tools/template/store/{name_cc}/index.js
