+++ build-tools/template/store/{name_cc}/{name_cc}.js
