+++ package.json
@@ -5,16 +5,16 @@
   "author": "Hendrik-Jan de Harder <hendrik-jan@mediamonks.com>",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --inline --progress --config build/webpack.dev.conf.js",
-    "dev-dhr": "cross-env NODE_ENV=development node build/dev-server.js --disableHotReloading",
-    "build": "cross-env NODE_ENV=production node build/build.js",
+    "dev": "cross-env NODE_ENV=development webpack-dev-server --inline --progress --config build-tools/config/webpack/webpack.dev.conf.js",
+    "dev-dhr": "cross-env NODE_ENV=development webpack-dev-server --inline --progress --config build-tools/config/webpack/webpack.dev.conf.js --disableHotReloading",
+    "build": "cross-env NODE_ENV=production node build-tools/script/build.js",
     "eslint": "eslint src --ext .js --cache",
     "tslint": "tslint src/**/*.ts -p tsconfig.json",
     "stylelint": "stylelint \"./src/**/*.scss\" --cache",
-    "preview": "node build/preview-server.js",
+    "preview": "node build-tools/script/preview-server.js",
     "prepush": "node bin/prePush.js",
     "precommit": "lint-staged",
-    "analyze": "webpack-bundle-analyzer stats.json ./dist",
+    "analyze": "webpack-bundle-analyzer stats.json ./build",
     "postcommit": "git reset",
     "prettify": "prettier --write \"src/**/*.{js,ts}\""
   },
@@ -76,7 +76,6 @@
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-prettier": "^2.1.2",
     "eventsource-polyfill": "^0.9.6",
-    "express": "^4.13.3",
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^1.1.5",
     "friendly-errors-webpack-plugin": "^1.1.3",
