--- template/page/{name_pc}/{name_pc}.vue
@@ -1,9 +1,0 @@
-<style src="./{{name_pc}}.scss" module lang="scss"></style>
-<script src="./{{name_pc}}.js"></script>
-
-<template>
-	<div>
-		<h2>{{name_pc}}</h2>
-	</div>
-</template>
-
